+++ apps/api/src/services/data-provider/data-provider.module.ts
@@ -1,6 +1,5 @@
 import { ConfigurationModule } from '@ghostfolio/api/services/configuration.module';
 import { CryptocurrencyModule } from '@ghostfolio/api/services/cryptocurrency/cryptocurrency.module';
-import { TrackinsightDataEnhancerService } from '@ghostfolio/api/services/data-provider/data-enhancer/trackinsight/trackinsight.service';
 import { GhostfolioScraperApiService } from '@ghostfolio/api/services/data-provider/ghostfolio-scraper-api/ghostfolio-scraper-api.service';
 import { RakutenRapidApiService } from '@ghostfolio/api/services/data-provider/rakuten-rapid-api/rakuten-rapid-api.service';
 import { YahooFinanceService } from '@ghostfolio/api/services/data-provider/yahoo-finance/yahoo-finance.service';
@@ -17,13 +16,7 @@ import { DataProviderService } from './data-provider.service';
     DataProviderService,
     GhostfolioScraperApiService,
     RakutenRapidApiService,
-    TrackinsightDataEnhancerService,
     YahooFinanceService,
-    {
-      inject: [TrackinsightDataEnhancerService],
-      provide: 'DataEnhancers',
-      useFactory: (trackinsight) => [trackinsight]
-    },
     {
       inject: [
         AlphaVantageService,